+++ include/um2/ray_casting/AngularQuadrature.hpp
@@ -68,9 +68,8 @@ setChebyshevAngularQuadrature(Size degree,
 
   // Angles
   angles.resize(degree);
-  T constexpr pi_v = pi<T>();
   T const degree_4 = static_cast<T>(degree) / static_cast<T>(4);
-  T const pi_deg = pi_v * degree_4; 
+  T const pi_deg = pi<T> * degree_4; 
   for (Size i = 0; i < degree; ++i) {
     angles[i] = pi_deg * static_cast<T>(2 * i - 1);
   }
