+++ include/um2/stdlib/String.hpp
@@ -99,7 +99,6 @@ public:
   PURE HOSTDEV [[nodiscard]] constexpr auto
   capacity() const noexcept -> uint64_t;
 
-  // cppcheck-suppress functionConst; justification: can't be const
   PURE HOSTDEV [[nodiscard]] constexpr auto
   data() noexcept -> char *;
 
@@ -184,7 +183,6 @@ public:
   getShortCap() noexcept -> uint64_t;
 
   HOSTDEV [[nodiscard]] HIDDEN constexpr auto
-  // cppcheck-suppress functionConst; justification: can't be const
   getLongPointer() noexcept -> char *;
 
   HOSTDEV [[nodiscard]] HIDDEN constexpr auto
