+++ include/um2/stdlib/Vector.hpp
@@ -44,7 +44,6 @@ public:
   HOSTDEV constexpr Vector(Vector && v) noexcept;
 
   // Initializer-list constructor should not be explicit
-  // cppcheck-suppress noExplicitConstructor; justified
   HOSTDEV constexpr Vector(std::initializer_list<T> const & list) noexcept;
 
   //==============================================================================
@@ -63,14 +62,12 @@ public:
   max_size() noexcept -> Size;
 
   PURE HOSTDEV [[nodiscard]] constexpr auto
-  // cppcheck-suppress functionConst; justification: cannot be const
   begin() noexcept -> T *;
 
   PURE HOSTDEV [[nodiscard]] constexpr auto
   begin() const noexcept -> T const *;
 
   PURE HOSTDEV [[nodiscard]] constexpr auto
-  // cppcheck-suppress functionConst; justification: cannot be const
   end() noexcept -> T *;
 
   PURE HOSTDEV [[nodiscard]] constexpr auto
@@ -92,21 +89,18 @@ public:
   cend() const noexcept -> T const *;
 
   PURE HOSTDEV [[nodiscard]] constexpr auto
-  // cppcheck-suppress functionConst; justification: cannot be const
   front() noexcept -> T &;
 
   PURE HOSTDEV [[nodiscard]] constexpr auto
   front() const noexcept -> T const &;
 
   PURE HOSTDEV [[nodiscard]] constexpr auto
-  // cppcheck-suppress functionConst; justification: cannot be const
   back() noexcept -> T &;
 
   PURE HOSTDEV [[nodiscard]] constexpr auto
   back() const noexcept -> T const &;
 
   PURE HOSTDEV [[nodiscard]] constexpr auto
-  // cppcheck-suppress functionConst; justification: cannot be const
   data() noexcept -> T *;
 
   PURE HOSTDEV [[nodiscard]] constexpr auto
@@ -136,7 +130,6 @@ public:
   //==============================================================================
 
   PURE HOSTDEV constexpr auto
-  // cppcheck-suppress functionConst; justification: cannot be const
   operator[](Size i) noexcept -> T &;
 
   PURE HOSTDEV constexpr auto
