+++ include/um2/stdlib/numeric.hpp
@@ -1,5 +1,24 @@
 #pragma once
 
-#include "./numeric/iota.hpp"
+#include <um2/config.hpp>
 
 #include <numeric>
+
+namespace um2
+{
+
+//==============================================================================
+// iota 
+//==============================================================================
+
+template <class T>
+HOSTDEV constexpr
+void iota(T * first, T * const last, T value)
+{
+  while (first != last) {
+    *first++ = value;
+    ++value;
+  }
+}
+
+} // namespace um2
