+++ tests/stdlib/memory/construct_at.cpp
@@ -1,4 +1,4 @@
-#include <um2/stdlib/memory/construct_at.hpp>
+#include <um2/stdlib/memory.hpp>
 
 #include <cstdlib>
